backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
bcrypt
python-multipart
pydantic
pyjwt